401エラー発生時のエラーハンドリングを追加
