+++ frontend/src/app/happiness/me/page.tsx
@@ -2,7 +2,7 @@
 import dynamic from 'next/dynamic'
 import { useState, useEffect, useContext } from 'react'
 import { useRouter } from 'next/navigation'
-import { useSession } from 'next-auth/react'
+import { signOut, useSession } from 'next-auth/react'
 import { Button, ButtonGroup, Grid } from '@mui/material'
 import { PeriodType } from '@/types/period'
 import { MessageType } from '@/types/message-type'
@@ -19,6 +19,7 @@ const BarGraph = dynamic(() => import('@/components/happiness/bar-graph'), {
 })
 import { myHappinessData } from '@/libs/graph'
 import { messageContext } from '@/contexts/message-context'
+import { ERROR_TYPE } from '@/libs/constants'
 import { fetchData } from '@/libs/fetch'
 import { toDateTime } from '@/libs/date-converter'
 import { useTokenFetchStatus } from '@/hooks/token-fetch-status'
@@ -60,10 +61,19 @@ const HappinessMe: React.FC = () => {
       setMyHappiness(myHappinessData(data))
     } catch (error) {
       console.error('Error fetching data:', error)
-      noticeMessageContext.showMessage(
-        '幸福度の検索に失敗しました',
-        MessageType.Error
-      )
+      if (error instanceof Error && error.message === ERROR_TYPE.UNAUTHORIZED) {
+        noticeMessageContext.showMessage(
+          '再ログインしてください',
+          MessageType.Error
+        )
+        signOut({ redirect: false })
+        router.push('/login')
+      } else {
+        noticeMessageContext.showMessage(
+          '幸福度の検索に失敗しました',
+          MessageType.Error
+        )
+      }
     }
   }
