+++ frontend/src/components/layout/index.tsx
@@ -2,19 +2,10 @@
 import { useEffect, useState } from 'react'
 import { usePathname } from 'next/navigation'
 import theme from '@/theme'
-import {
-  Box,
-  CssBaseline,
-  Toolbar,
-  Snackbar,
-  Alert,
-  ThemeProvider,
-} from '@mui/material'
+import { Box, CssBaseline, Toolbar, ThemeProvider } from '@mui/material'
 
 import Header from '@/components/header'
 import Sidebar from '@/components/sidebar/sidebar'
-import { messageContext } from '@/contexts/message-context'
-import { useNoticeMessage } from '@/hooks/notice-message'
 import { SessionProvider } from 'next-auth/react'
 
 interface LayoutProps {
@@ -23,7 +14,6 @@ interface LayoutProps {
 }
 
 const Layout: React.FC<LayoutProps> = ({ simple = false, children }) => {
-  const noticeMessageContext = useNoticeMessage()
   const [isOpen, setIsOpen] = useState(false)
   const pathname = usePathname()
 
@@ -44,45 +34,26 @@ const Layout: React.FC<LayoutProps> = ({ simple = false, children }) => {
           </Box>
         </Box>
       ) : (
-        <messageContext.Provider value={noticeMessageContext}>
-          <SessionProvider>
-            <Box sx={{ display: 'flex' }}>
-              <CssBaseline />
-              <Header
-                handleDrawerOpen={() => {
-                  setIsOpen(true)
-                }}
-              />
-              <Box sx={{ width: 1 }}>
-                <Toolbar />
-                {children}
-              </Box>
-              <Sidebar
-                isOpen={isOpen}
-                handleDrawerClose={() => {
-                  setIsOpen(false)
-                }}
-              />
+        <SessionProvider>
+          <Box sx={{ display: 'flex' }}>
+            <CssBaseline />
+            <Header
+              handleDrawerOpen={() => {
+                setIsOpen(true)
+              }}
+            />
+            <Box sx={{ width: 1 }}>
+              <Toolbar />
+              {children}
             </Box>
-          </SessionProvider>
-          {noticeMessageContext.message && (
-            <Snackbar
-              open={true}
-              autoHideDuration={6000}
-              onClose={noticeMessageContext.clearMessage}
-              anchorOrigin={{ vertical: 'bottom', horizontal: 'right' }}
-            >
-              <Alert
-                onClose={noticeMessageContext.clearMessage}
-                severity={noticeMessageContext.type}
-                variant="filled"
-                sx={{ width: '100%' }}
-              >
-                {noticeMessageContext.message}
-              </Alert>
-            </Snackbar>
-          )}
-        </messageContext.Provider>
+            <Sidebar
+              isOpen={isOpen}
+              handleDrawerClose={() => {
+                setIsOpen(false)
+              }}
+            />
+          </Box>
+        </SessionProvider>
       )}
     </ThemeProvider>
   )
