+++ frontend/src/libs/constants.ts
@@ -2,3 +2,7 @@ export const PROFILE_TYPE = {
   GENERAL: 'general',
   ADMIN: 'admin',
 }
+
+export const ERROR_TYPE = {
+  UNAUTHORIZED: 'Unauthorized',
+}
